requirements.txt
flask
requests
solders
base58
python-dotenv